third_party/WebKit/LayoutTests/fast/scroll-behavior/scroll-into-view-scrolls-layout-viewport.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
  function runTest() {
    test( () => {
      assert_true(typeof(window.internals) !== 'undefined',
                  "This test requires window.internals");

      internals.setPageScaleFactor(2.0);
      const target = document.getElementById("box");

      window.scrollTo(0, 0);
      internals.setVisualViewportOffset(0, 0);

      target.scrollIntoView();
      assert_equals(window.visualViewport.pageLeft, window.scrollX);
      assert_equals(window.visualViewport.pageTop, window.scrollY);
      assert_equals(window.scrollX, 0);
      assert_greater_than(window.scrollY, 0);

      window.scrollTo(0, 0);
      internals.setVisualViewportOffset(0, 0);

      target.scrollIntoViewIfNeeded();
      assert_equals(window.visualViewport.pageLeft, window.scrollX);
      assert_equals(window.visualViewport.pageTop, window.scrollY);
      assert_equals(window.scrollX, 0);
      assert_greater_than(window.scrollY, 0);
    }, "Test that element.scrollIntoView() scrolls only the layout viewport.");
  }

  onload = runTest;
</script>

<style>
  body {
    height: 800px;
    width: 800px;
  }
  #box {
    background-color: black;
    top: 700px;
    left: 100px;
    height: 40px;
    width: 40px;
    position: absolute;
  }
</style>

<div id="box"></div>
